AE2 Finished

+++ PlaceMyBet_1/PlaceMyBet_1/Models/Apuesta.cs
@@ -7,7 +7,7 @@ namespace PlaceMyBet_1.Models
 {
     public class Apuesta
     {
-        public Apuesta(int apuesta_id, int mercado_ref, string usuario_ref, int dinero)
+        public Apuesta(int apuesta_id, int mercado_ref, string usuario_ref, double dinero)
         {
             this.apuesta_id = apuesta_id;
             this.mercado_ref = mercado_ref;
@@ -18,7 +18,7 @@ namespace PlaceMyBet_1.Models
         public int apuesta_id { get; set; }
         public int mercado_ref { get; set; }
         public string usuario_ref { get; set; }
-        public int dinero { get; set; }
+        public double dinero { get; set; }
 
     }
 }
