+++ PlaceMyBet_1/PlaceMyBet_1/Models/MercadosRepository.cs
@@ -14,7 +14,7 @@ namespace PlaceMyBet_1.Models
             List<ArrayList> datos = Common.MasterRepository.Leer("SELECT * FROM mercados");
             foreach (var item in datos)
             {
-                mercados.Add(new Mercado((int)item[0], (double)item[1], (double)item[2], (double)item[3], (double)item[4], item[5].ToString()));
+                mercados.Add(new Mercado((int)item[0], (float)item[1], (float)item[2], (float)item[3], (float       )item[4], item[5].ToString()));
             }
             return mercados;
         }
